move fetchData to utils folder
